Refactor ui/components/includes
