+++ app/imports/ui/components/includes/card-changelog/index.js
@@ -1,0 +1,2 @@
+import './card-changelog.html';
+import './card-changelog.js';
