+++ app/imports/ui/components/includes/clear-field-button/clear-field-button.html
