+++ app/imports/ui/components/includes/clear-field-button/index.js
@@ -1,0 +1,1 @@
+import './clear-field-button.html';
