+++ app/imports/ui/components/includes/clearable-field/clear-field/clear-field.js
