+++ app/imports/ui/components/includes/clearable-field/clear-field/index.js
@@ -1,0 +1,2 @@
+import './clear-field.html';
+import './clear-field.js';
