+++ app/imports/ui/components/includes/clearable-field/clearable-field.js
