+++ app/imports/ui/components/includes/clearable-field/index.js
@@ -1,0 +1,4 @@
+import './clearable-field.html';
+import './clearable-field.js';
+
+import './clear-field';
