+++ app/imports/ui/components/includes/datepicker/datepicker.html
