+++ app/imports/ui/components/includes/datepicker/index.js
@@ -1,0 +1,2 @@
+import './datepicker.html';
+import './datepicker.js';
