+++ app/imports/ui/components/includes/delete-button/delete-button.html
