+++ app/imports/ui/components/includes/delete-button/index.js
@@ -1,0 +1,2 @@
+import './delete-button.html';
+import './delete-button.js';
