+++ app/imports/ui/components/includes/file-item/file-item.js
