+++ app/imports/ui/components/includes/file-item/index.js
@@ -1,0 +1,2 @@
+import './file-item.html';
+import './file-item.js';
