+++ app/imports/ui/components/includes/file-uploader/index.js
@@ -1,0 +1,2 @@
+import './file-uploader.html';
+import './file-uploader.js';
