+++ app/imports/ui/components/includes/index.js
@@ -1,31 +1,17 @@
-import './simple-card.html';
-import './simple-card.js';
-import './list-item.html';
-import './list-item.js';
-import './list-sub-item.html';
-import './list-sub-item.js';
-import './card-changelog.html';
-import './card-changelog.js';
-import './time-picker.html';
-import './preloader.html';
-import './time-picker.js';
-import './clear-field-button.html';
-import './datepicker.html';
-import './datepicker.js';
-import './file-uploader.html';
-import './file-uploader.js';
-import './file-uploader-wrapper';
-import './modal-delete-button.html';
-import './modal-delete-button.js';
-import './clear-field.html';
-import './clear-field.js';
-import './file-item.html';
-import './file-item.js';
-import './clearable-field.html';
-import './clearable-field.js';
-import './fields';
-import './delete-button.html';
-import './delete-button.js';
-import './subcards';
 import './card';
+import './card-changelog';
+import './clear-field-button';
+import './clearable-field';
+import './datepicker';
+import './delete-button';
+import './fields';
+import './file-item';
+import './file-uploader';
+import './file-uploader-wrapper';
+import './list-item';
+import './modal-delete-button';
+import './preloader';
 import './score-table';
+import './simple-card';
+import './subcards';
+import './time-picker';
