+++ app/imports/ui/components/includes/list-item/index.js
@@ -1,0 +1,4 @@
+import './list-item.html';
+import './list-item.js';
+
+import './list-sub-item';
