+++ app/imports/ui/components/includes/list-item/list-sub-item/index.js
@@ -1,0 +1,2 @@
+import './list-sub-item.html';
+import './list-sub-item.js';
