+++ app/imports/ui/components/includes/modal-delete-button/index.js
@@ -1,0 +1,2 @@
+import './modal-delete-button.html';
+import './modal-delete-button.js';
