+++ app/imports/ui/components/includes/modal-delete-button/modal-delete-button.js
