+++ app/imports/ui/components/includes/preloader/index.js
@@ -1,0 +1,1 @@
+import './preloader.html';
