+++ app/imports/ui/components/includes/simple-card/index.js
@@ -1,0 +1,2 @@
+import './simple-card.html';
+import './simple-card.js';
