+++ app/imports/ui/components/includes/simple-card/simple-card.js
