+++ app/imports/ui/components/includes/time-picker/index.js
@@ -1,0 +1,2 @@
+import './time-picker.html';
+import './time-picker.js';
